server/requirements.txt
fastapi
bcrypt
PyJWT